Function streams_orderednogaps::ordered_no_gaps

pub fn ordered_no_gaps<S, F, K>(
    stream: S, 
    zero: K, 
    key: F
) -> OrderedNoGaps<S, F, K> where
    S: Stream,
    S::Item: Ord,
    F: Fn(&S::Item) -> K,
    K: Successor + PartialEq, 

Given a function to retrieve the key of each value, the zero key value and an underlying stream.

Guarantees order of emitted values from the provided stream by the specified key and that no key values are skipped. Halts until the next expected key is received.

WARNING: The system's memory is an implicit upper bound on how much this combinator can buffer internally.